Add files via upload
Name: Steven Kuzhipala
Part of #28

In this notebook, I have reviewed if there is any bias in the given data set for pneumonia, by checking using matplotlib the class sizes for both the classes. After doing so, there was a clear difference between the number of images available for the NORMAL class and less for the pneumonia class. That means that during the training process, the model could be biased in its classification as well, simply because of the data we train it with.

To solve this problem, I decided to try to augment the number of images in the pneumonia class along with images that are simply resized and are in the same color format as the initial data set, to not cause any other differences for identification. I made sure that the number of dimensions or channels in each of the images were 3, including R, G and B so that the images are in grayscale not causing a distinct change in one of the classes.

To perform the rescalign and sizing I used the keras model.